Turn a nested model into a tidy tibble

Use broom functions on fitted nested models.

tidy.nested_model_fit summarises components of each model within a nested model fit, indicating which nested data frame each row corresponds to.

glance.nested_model_fit summarises a nested model, returning a tibble::tibble() with 1 row.

glance_nested summarises each model within a nested model fit, returning a tibble::tibble() with the same number of rows as the number of inner models.

Usage

# S3 method for nested_model_fit
tidy(x, ...)

# S3 method for nested_model_fit
glance(x, ...)

glance_nested(x, ...)

Arguments

x

The object to be converted into a tidy tibble::tibble().

...

Additional arguments passed into their respective functions. (e.g. for tidy.nested_model_fit, parsnip::tidy.model_fit()).

Value

A tibble::tibble(). With glance.nested_model_fit(), the tibble will have 1 row.

Details

generics::glance() states that glance methods should always return 1 row outputs for non-empty inputs. The 'nestedmodels' package is no exception: glance methods will combine rows to produce a result with a single row. Specifically:

If a column contains 1 unique value, that value is used.

If a column is numeric, the mean will be calculated.

Otherwise, the results will be combined into a list.

See also

Examples

model <- parsnip::linear_reg() %>%
  parsnip::set_engine("lm") %>%
  nested()
  
fit <- fit(model, z ~ ., tidyr::group_by(example_nested_data, .data$id))
#> Error: 'group_by' is not an exported object from 'namespace:tidyr'

tidy(fit)
#> Error in tidy(fit): could not find function "tidy"
glance(fit)
#> Error in glance(fit): could not find function "glance"
glance_nested(fit)
#> Error in glance_nested(fit): `x` must have class: <nested_model_fit>.
#> ℹ Actual class: <function>.