Nest transformation

step_nest creates a specification of a recipe step that will transform data using tidyr::nest()

Usage

step_nest(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  names = NULL,
  lookup_table = NULL,
  skip = FALSE,
  id = recipes::rand_id("nest")
)

# S3 method for step_nest
prep(x, training, info = NULL)

# S3 method for step_nest
bake(object, new_data, ...)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

...

One or more selector functions to choose variables. For step_nest, this indicates the variables which will not be nested. See recipes::selections() for more details.

role

Not used by this step since the new variables are assigned a custom role.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

skip

A logical. Should the step be skipped when the recipe is baked by bake()? While all operations are baked when prep() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id

A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new step added to the sequence of any existing operations.

Details

Both the columns to nest and the outer columns must be included in the recipe spec. step_nest will give the non-nested columns a new role: 'nested_id'. This should not be altered. The actual transformation applied to the data resembles the following: tidyr::nest(data, data = c(...)) This means that you can specify columns to nest in the traditional c(cols) way, or by not wrapping the columns in c(), since the function does it automatically.

If you nest the data yourself, the resulting nested column must be named 'data'. For this reason, you must not have a column named 'data' in your inputted data. If you plan to use recipes::prep() and recipes::bake() on your recipe, make sure your step_nest step is last, since other recipe steps will not be able to evaluate on nested data.

Examples

data("example_nested_data")

recipe <- recipes::recipe(example_nested_data, z ~ x + id) %>%
  step_nest(-id) # equivalent to tidyr::nest(example_nested_data, data = -id)

data <- recipe %>%
  recipes::prep() %>%
  recipes::bake(NULL)

recipe2 <- recipes::recipe(example_nested_data, z ~ x + id) %>%
  step_nest(x, z) 
# equivalent to tidyr::nest(example_nested_data, data = c(x,z))

data2 <- recipe %>%
  recipes::prep() %>%
  recipes::bake(NULL)
  
identical(data1, data2) # TRUE
#> Error in identical(data1, data2): object 'data1' not found