Turn a nested model into a tidy tibble

Use broom functions on fitted nested models.

tidy.nested_model_fit summarises components of each model within a nested model fit, indicating which nested data frame each row corresponds to.

glance.nested_model_fit summarises a nested model, returning a tibble::tibble() with 1 row.

glance_nested summarises each model within a nested model fit, returning a tibble::tibble() with the same number of rows as the number of inner models.

Usage

# S3 method for nested_model_fit
tidy(x, ...)

# S3 method for nested_model_fit
glance(x, ...)

glance_nested(x, ...)

Arguments

x

The object to be converted into a tidy tibble::tibble().

...

Additional arguments passed into their respective functions. (e.g. for tidy.nested_model_fit, parsnip::tidy.model_fit()).

Value

A tibble::tibble(). With glance.nested_model_fit(), the tibble will have 1 row.

Details

generics::glance() states that glance methods should always return 1 row outputs for non-empty inputs. The 'nestedmodels' package is no exception: glance methods will combine rows to produce a result with a single row. Specifically:

If a column contains 1 unique value, that value is used.

If a column is numeric, the mean will be calculated.

Otherwise, the results will be combined into a list.

See also

Examples

if (require("broom")) {
  model <- parsnip::linear_reg() %>%
    parsnip::set_engine("lm") %>%
    nested()

  fit <- fit(model, z ~ x + y + a + b, 
             dplyr::group_by(example_nested_data, id))

  tidy(fit)
  glance(fit)
  glance_nested(fit)
}
#> Loading required package: broom
#> # A tibble: 20 × 13
#>       id r.squ…¹ adj.r…² sigma stati…³  p.value    df logLik   AIC   BIC devia…⁴
#>    <int>   <dbl>   <dbl> <dbl>   <dbl>    <dbl> <dbl>  <dbl> <dbl> <dbl>   <dbl>
#>  1     1   0.453   0.404 18.0     9.31 1.43e- 5     4  -213.  438.  449.  14617.
#>  2     2   0.499   0.454 15.3    11.2  2.18e- 6     4  -205.  421.  433.  10476.
#>  3     3   0.456   0.408 12.9     9.44 1.25e- 5     4  -196.  404.  416.   7477.
#>  4     4   0.594   0.558  7.25   16.5  2.23e- 8     4  -167.  347.  358.   2368.
#>  5     5   0.293   0.230 18.5     4.65 3.15e- 3     4  -214.  440.  452.  15397.
#>  6     6   0.864   0.852  4.83   71.7  6.12e-19     4  -147.  306.  318.   1051.
#>  7     7   0.477   0.430 14.9    10.2  5.52e- 6     4  -203.  418.  430.   9932.
#>  8     8   0.779   0.759 12.9    39.6  3.42e-14     4  -196.  405.  416.   7535.
#>  9     9   0.534   0.493 14.2    12.9  4.40e- 7     4  -201.  414.  425.   9021.
#> 10    10   0.758   0.737  7.91   35.3  2.43e-13     4  -172.  355.  367.   2818.
#> 11    11   0.745   0.722  7.35   32.9  7.88e-13     4  -168.  348.  360.   2429.
#> 12    12   0.327   0.267 17.6     5.47 1.12e- 3     4  -212.  435.  447.  13883.
#> 13    13   0.485   0.439 17.5    10.6  3.93e- 6     4  -211.  435.  446.  13822.
#> 14    14   0.501   0.457 24.6    11.3  1.96e- 6     4  -228.  469.  480.  27142.
#> 15    15   0.215   0.146 21.9     3.09 2.49e- 2     4  -223.  457.  469.  21605.
#> 16    16   0.247   0.181 16.5     3.70 1.09e- 2     4  -208.  429.  440.  12204.
#> 17    17   0.778   0.759  6.16   39.5  3.48e-14     4  -159.  330.  342.   1709.
#> 18    18   0.876   0.865 10.3    79.6  7.99e-20     4  -185.  382.  393.   4791.
#> 19    19   0.536   0.495 11.1    13.0  4.10e- 7     4  -189.  389.  400.   5510.
#> 20    20   0.611   0.576  9.49   17.6  8.93e- 9     4  -181.  374.  385.   4057.
#> # … with 2 more variables: df.residual <int>, nobs <int>, and abbreviated
#> #   variable names ¹​r.squared, ²​adj.r.squared, ³​statistic, ⁴​deviance