Turn a nested model into a tidy tibble

Use broom functions on fitted nested models.

tidy.nested_model_fit() summarises components of each model within a nested model fit, indicating which nested data frame each row corresponds to.

glance.nested_model_fit() summarises a nested model, returning a tibble::tibble() with 1 row.

glance_nested() summarises each model within a nested model fit, returning a tibble::tibble() with the same number of rows as the number of inner models.

Usage

# S3 method for nested_model_fit
tidy(x, ...)

# S3 method for nested_model_fit
glance(x, ...)

glance_nested(x, ...)

Arguments

x

A nested_model_fit object produced by fit.nested_model().

...

Additional arguments passed into their respective functions. (e.g. for tidy.nested_model_fit(), parsnip::tidy.model_fit()).

Value

A tibble::tibble(). With glance.nested_model_fit(), the tibble will have 1 row.

Details

generics::glance() states that glance() methods should always return 1 row outputs for non-empty inputs. The 'nestedmodels' package is no exception: glance() methods will combine rows to produce a result with a single row. Specifically:

If a column contains 1 unique value, that value is used.

If a column is numeric, the mean will be calculated.

Otherwise, the results will be combined into a list.

See also

Examples

if (require("broom")) {
  data <- dplyr::filter(example_nested_data, id %in% 1:5)

  model <- parsnip::linear_reg() %>%
    parsnip::set_engine("lm") %>%
    nested()

  fit <- fit(
    model, z ~ x + y + a + b,
    dplyr::group_by(data, id)
  )

  tidy(fit)
  glance(fit)
  glance_nested(fit)
}
#> Loading required package: broom
#> # A tibble: 5 × 13
#>      id r.squared adj.r…¹ sigma stati…² p.value    df logLik   AIC   BIC devia…³
#>   <int>     <dbl>   <dbl> <dbl>   <dbl>   <dbl> <dbl>  <dbl> <dbl> <dbl>   <dbl>
#> 1     1    0.161  0.0861  18.0     2.15 8.95e-2     4  -213.  438.  449.  14617.
#> 2     2    0.359  0.302   15.3     6.30 4.12e-4     4  -205.  421.  433.  10476.
#> 3     3    0.218  0.149   12.9     3.14 2.31e-2     4  -196.  404.  416.   7477.
#> 4     4    0.0867 0.00554  7.25    1.07 3.83e-1     4  -167.  347.  358.   2368.
#> 5     5    0.0934 0.0128  18.5     1.16 3.42e-1     4  -214.  440.  452.  15397.
#> # … with 2 more variables: df.residual <int>, nobs <int>, and abbreviated
#> #   variable names ¹​adj.r.squared, ²​statistic, ³​deviance